src/third_party/blink/web_tests/fast/css/forced-colors-mode/forced-colors-mode-03.html

<!DOCTYPE html>
<html>
<head>
<title>Forced colors mode - highlighted text.</title>
<style type="text/css">
  #a {
    color: orange;
    forced-color-adjust: auto;
  }
  #b {
    color: orange;
    forced-color-adjust: none;
  }
</style>
</head>
<body>
  <mark id="a">
    This text should be black in forced colors mode because forced-color-adjust
    is auto. The backgroud color should be yellow because the default
    highlighted colors should not be overridden in forced colors mode.
  </mark>
  <mark id="b">
    This text should be orange in forced colors mode because
    forced-color-adjust is none. The backgroud color should be yellow because
    the default highlighted colors should not be overridden in forced colors
    mode.
  </mark>
</body>

<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
  test(function(){
    assert_equals(getComputedStyle(a).color, "rgb(0, 0, 0)");

    assert_equals(getComputedStyle(b).color, "rgb(255, 165, 0)");

    assert_equals(getComputedStyle(a).backgroundColor, "rgb(255, 255, 0)");

    assert_equals(getComputedStyle(b).backgroundColor, "rgb(255, 255, 0)");

  }, "Checks that default highlighted text style does not get overridden in forced colors mode.");
</script>
</html>
